# services/product-catalog-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manages products, SKUs, categories, and images for ZapMarket.",
        "title": "Product Catalog Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/v1/skus": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skus"
                ],
                "summary": "List SKUs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by product UUID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by SKU code (partial match)",
                        "name": "sku_code",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by active status",
                        "name": "is_active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.SKU"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skus"
                ],
                "summary": "Create a SKU",
                "parameters": [
                    {
                        "description": "SKU payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateSKURequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.SKU"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/skus/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skus"
                ],
                "summary": "Get SKU by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SKU UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.SKU"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skus"
                ],
                "summary": "Update SKU",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SKU UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SKU payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateSKURequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.SKU"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "skus"
                ],
                "summary": "Delete SKU",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SKU UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.SKU": {
            "type": "object",
            "properties": {
                "compare_price": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "price_amount": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "sku_code": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "variant_attributes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "weight_grams": {
                    "type": "integer"
                }
            }
        },
        "github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/http.ErrorInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "http.CreateSKURequest": {
            "type": "object",
            "properties": {
                "compare_price": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "price_amount": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "sku_code": {
                    "type": "string"
                },
                "variant_attrs": {},
                "weight_grams": {
                    "type": "integer"
                }
            }
        },
        "http.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "http.UpdateSKURequest": {
            "type": "object",
            "properties": {
                "compare_price": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "price_amount": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "sku_code": {
                    "type": "string"
                },
                "variant_attrs": {},
                "weight_grams": {
                    "type": "integer"
                }
            }
        }
    }
}